+++ src/registry/removeWorkerObserverFromRegistry.ts
@@ -1,5 +1,9 @@
-import { TDreamstateWorker, TUpdateObserver } from "../types";
-import { CONTEXT_OBSERVERS_REGISTRY } from "../internals";
+import type { ContextWorker } from "../management";
+import type { TDreamstateWorker, TUpdateObserver } from "../types";
+import {
+  CONTEXT_OBSERVERS_REGISTRY,
+  CONTEXT_WORKERS_REGISTRY
+} from "../internals";
 import { unRegisterWorker } from "./unRegisterWorker";
 
 import { log } from "../../build/macroses/log.macro";
@@ -11,12 +15,22 @@ export function removeWorkerObserverFromRegistry(
   workerConstructor: TDreamstateWorker,
   observer: TUpdateObserver
 ): void {
-  // Remove observer.
   CONTEXT_OBSERVERS_REGISTRY.get(workerConstructor)!.delete(observer);
 
   log.info("Worker observer removed:", workerConstructor.name);
 
   if (CONTEXT_OBSERVERS_REGISTRY.get(workerConstructor)!.size === 0) {
-    unRegisterWorker(workerConstructor);
+    const instance: ContextWorker | undefined = CONTEXT_WORKERS_REGISTRY.get(workerConstructor)!;
+
+    if (instance) {
+      instance["onProvisionEnded"]();
+
+      log.info("Worker provision ended:", workerConstructor.name);
+
+      unRegisterWorker(workerConstructor);
+    } else {
+      log.error("Worker failed to unregister:", workerConstructor.name);
+      throw new Error("Could not find manager instance when provision ended.");
+    }
   }
 }